+++ QUANTAXIS/QABacktest/QABacktest.py
@@ -81,7 +81,7 @@ class QA_Backtest():
         self.account = None
         self.portfolio = None
 
-        #🛠todo market_type 应该放在 QA_Market对象里的一个属性
+        # 🛠todo market_type 应该放在 QA_Market对象里的一个属性
         self.market = QA_Market()
         self.market_type = market_type
 
@@ -93,8 +93,8 @@ class QA_Backtest():
         self.end = end
         self.code_list = code_list
 
-        #🛠todo 检查start日期和结束end日期是否正确
-        #🛠todo 检查code list 是否合法
+        # 🛠todo 检查start日期和结束end日期是否正确
+        # 🛠todo 检查code list 是否合法
 
         # 根据 市场类型，回测周期频率， 和股票代码列表 获取回测数据
         if self.market_type is MARKET_TYPE.STOCK_CN and self.frequence is FREQUENCE.DAY:
@@ -136,7 +136,7 @@ class QA_Backtest():
         # 如果出现了日期的改变 才会进行结算的事件
         _date = None
         for data in self.ingest_data:  # 对于在ingest_data中的数据
-            #<class 'QUANTAXIS.QAData.QADataStruct.QA_DataStruct_Stock_day'>
+            # <class 'QUANTAXIS.QAData.QADataStruct.QA_DataStruct_Stock_day'>
             date = data.date[0]
             if self.market_type is MARKET_TYPE.STOCK_CN:  # 如果是股票市场
                 if _date != date:  # 如果新的date
@@ -145,7 +145,7 @@ class QA_Backtest():
                     # 往 broker 和 account 发送 settle 事件
                     try:
                         self.market.trade_engine.join()
-                        #time.sleep(2)
+                        # time.sleep(2)
                         self.market._settle(self.broker_name)
 
                     except Exception as e: