+++ QUANTAXIS/QAMarket/QAMarket.py
@@ -24,8 +24,9 @@
 
 
 import datetime
-
+import time
 import numpy as np
+import sched
 
 from QUANTAXIS.QAARP.QAAccount import QA_Account
 from QUANTAXIS.QAEngine.QAEvent import QA_Event
@@ -38,7 +39,7 @@ from QUANTAXIS.QAMarket.QAShipaneBroker import QA_SPEBroker
 from QUANTAXIS.QAMarket.QASimulatedBroker import QA_SimulatedBroker
 from QUANTAXIS.QAMarket.QATrade import QA_Trade
 from QUANTAXIS.QAUtil.QALogs import QA_util_log_info
-from QUANTAXIS.QAUtil.QAParameter import (ACCOUNT_EVENT, AMOUNT_MODEL,
+from QUANTAXIS.QAUtil.QAParameter import (ACCOUNT_EVENT, AMOUNT_MODEL, ORDER_STATUS,
                                           BROKER_EVENT, BROKER_TYPE,
                                           ENGINE_EVENT, FREQUENCE,
                                           MARKET_EVENT, ORDER_EVENT,
@@ -61,6 +62,8 @@ class QA_Market(QA_Trade):
 
         Keyword Arguments:
             if_start_orderthreading {bool} -- 是否在初始化的时候开启查询子线程(实盘需要) (default: {False})
+
+        @2018-08-06 change : 子线程全部变成后台线程 market线程崩了 子线程全部结束
         """
 
         super().__init__()
@@ -78,7 +81,7 @@ class QA_Market(QA_Trade):
         self.running_time = None
         self.last_query_data = None
         self.if_start_orderthreading = if_start_orderthreading
-        self.order_handler = False
+        self.order_handler = QA_OrderHandler()
 
     def __repr__(self):
         '''
@@ -100,7 +103,7 @@ class QA_Market(QA_Trade):
         self.running_time = data.datetime[0]
         for item in self.session.values():
             # session里面是已经注册的account
-            self.event_queue.put(QA_Task(
+            self.submit(QA_Task(
                 worker=item,  # item 是Account 类型， 是 QA_Work类型， 处理这个 事件
                 event=QA_Event(
                     event_type=ENGINE_EVENT.UPCOMING_DATA,
@@ -114,6 +117,30 @@ class QA_Market(QA_Trade):
                     query_trade=self.query_trade
                 )
             ))
+
+    def submit(self, QATask, nowait=False):
+        """submit 一个任务给QAMarket的event_queue
+
+        Arguments:
+            QATask {[type]} -- [description]
+
+        QATask 需要有
+            - worker (需要这个类继承了QA_Worker)
+            - engine(默认qamarket所在的thread)
+            - event - QA_Event
+                        - event_type
+                        - 自定义参数
+                        - callback
+
+        Keyword Arguments:
+            nowait {bool} -- [description] (default: {False})
+        """
+
+        assert isinstance(QATask, QA_Task)
+        if nowait:
+            self.event_queue.put_nowait(QATask)
+        else:
+            self.submit(QATask)
 
     def start(self):
         self.trade_engine.start()
@@ -129,7 +156,8 @@ class QA_Market(QA_Trade):
         if broker in self._broker.keys():
 
             self.broker[broker] = self._broker[broker]()  # 在这里实例化
-            self.trade_engine.create_kernel('{}'.format(broker))
+            # 2018-08-06 change : 子线程全部变成后台线程 market线程崩了 子线程全部结束
+            self.trade_engine.create_kernel('{}'.format(broker), daemon=True)
             self.trade_engine.start_kernel('{}'.format(broker))
             # 开启trade事件子线程
             return True
@@ -139,7 +167,8 @@ class QA_Market(QA_Trade):
     def register(self, broker_name, broker):
         if broker_name not in self._broker.keys():
             self.broker[broker_name] = broker
-            self.trade_engine.create_kernel('{}'.format(broker_name))
+            self.trade_engine.create_kernel(
+                '{}'.format(broker_name), daemon=True)
             self.trade_engine.start_kernel('{}'.format(broker_name))
             return True
         else:
@@ -150,9 +179,11 @@ class QA_Market(QA_Trade):
         """
 
         self.if_start_orderthreading = True
-        self.order_handler = QA_OrderHandler()
-        self.trade_engine.create_kernel('ORDER')
+
+        self.order_handler.if_start_orderquery = True
+        self.trade_engine.create_kernel('ORDER', daemon=True)
         self.trade_engine.start_kernel('ORDER')
+        # self._update_orders()
 
     def get_account(self, account_cookie):
         return self.session[account_cookie]
@@ -201,7 +232,7 @@ class QA_Market(QA_Trade):
 
         Arguments:
             broker_id {[type]} -- [description]
-            account_id {[type]} -- [description]
+            account_cookie {[type]} -- [description]
         """
         try:
             if isinstance(self.broker[broker_name], QA_BacktestBroker):
@@ -226,7 +257,7 @@ class QA_Market(QA_Trade):
     def get_account_id(self):
         return list(self.session.keys())
 
-    def insert_order(self, account_id, amount, amount_model, time, code, price, order_model, towards, market_type, frequence, broker_name, money=None):
+    def insert_order(self, account_cookie, amount, amount_model, time, code, price, order_model, towards, market_type, frequence, broker_name, money=None):
         #strDbg = QA_util_random_with_topic("QA_Market.insert_order")
         #print(">-----------------------insert_order----------------------------->", strDbg)
 
@@ -278,7 +309,7 @@ class QA_Market(QA_Trade):
             # if price > self.last_query_data[0][2] or price < self.last_query_data[0][3]:
             flag = True
         if flag:
-            order = self.get_account(account_id).send_order(
+            order = self.get_account(account_cookie).send_order(
                 amount=amount, amount_model=amount_model, time=time, code=code, price=price,
                 order_model=order_model, towards=towards, money=money)
 
@@ -289,44 +320,125 @@ class QA_Market(QA_Trade):
             # print(datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S'))  # 日期格式化
             # print("<-------------------------------------------------")
             if order:
-                self.event_queue.put_nowait(
+                # print(order)
+                self.submit(
                     QA_Task(
-                        worker=self.broker[self.get_account(
-                            account_id).broker],
-                        engine=self.get_account(account_id).broker,
+                        worker=self.order_handler,
+                        engine='ORDER',
                         event=QA_Event(
+                            broker=self.broker[self.get_account(
+                                account_cookie).broker],
                             event_type=BROKER_EVENT.RECEIVE_ORDER,
                             order=order,
-                            callback=self.on_insert_order)))
+                            callback=self.on_insert_order)), nowait=True)
         else:
             pass
 
         #print("<-----------------------insert_order-----------------------------<", strDbg)
 
     def on_insert_order(self, order):
+        print('on_insert_order')
         print(order)
-        pass
+        print(order.status)
+        if order.status == ORDER_STATUS.FAILED:
+            """如果订单创建失败, 恢复状态
+
+            如果是买入单  恢复金钱 money
+
+            如果是卖出单  恢复股数 sell_available
+            """
+
+            self.session[order.account_cookie].cancel_order(order)
 
     def _renew_account(self):
         for account in self.session.values():
-            self.event_queue.put(
+            self.submit(
                 QA_Task(
                     worker=account,
                     event=QA_Event(
                         event_type=ACCOUNT_EVENT.SETTLE)))
 
-    def query_order(self, broker_name, order_id):
-
-        res = self.event_queue.put(
+    def _sync_position(self):
+        self.submit(
             QA_Task(
-                worker=self.broker[broker_name],
-                engine=broker_name,
+                worker=self.order_handler,
+                engine='ORDER',
                 event=QA_Event(
-                    order_id=order_id
+                    event_type=MARKET_EVENT.QUERY_POSITION,
+                    account_cookie=list(self.session.keys()),
+                    broker=[self.broker[item.broker]
+                            for item in self.session.values()]
                 )
-            ))
+            ), nowait=True
+        )
 
-        return res
+    def _sync_deals(self):
+        self.submit(
+            QA_Task(
+                worker=self.order_handler,
+                engine='ORDER',
+                event=QA_Event(
+                    event_type=MARKET_EVENT.QUERY_DEAL,
+                    account_cookie=list(self.session.keys()),
+                    broker=[self.broker[item.broker]
+                            for item in self.session.values()],
+                    event_queue=self.trade_engine.kernels_dict['ORDER'].queue
+                )
+            ), nowait=True
+        )
+
+    def _sync_orders(self):
+        self.submit(
+            QA_Task(
+                worker=self.order_handler,
+                engine='ORDER',
+                event=QA_Event(
+                    event_type=MARKET_EVENT.QUERY_ORDER,
+                    account_cookie=list(self.session.keys()),
+                    broker=[self.broker[item.broker]
+                            for item in self.session.values()],
+                    # 注意: 一定要给子线程的队列@@@!!!
+                    # 2018-08-08 yutiansut
+                    # 这个callback实现了子线程方法的自我驱动和异步任务
+                    event_queue=self.trade_engine.kernels_dict['ORDER'].queue
+                )
+            ), nowait=True
+        )
+
+    def sync_strategy(self, broker_name, account_cookie):
+        """同步  账户/委托/成交
+
+        Arguments:
+            broker_name {[type]} -- [description]
+            account_cookie {[type]} -- [description]
+        """
+        pass
+
+    def cancel_order(self):
+        pass
+
+    def cancel_all(self, broker_name, account_cookie):
+        try:
+            self.broker[broker_name].cancel_all(account_cookie)
+        except Exception as e:
+            print(e)
+
+    def query_order(self, account_cookie, realorder_id):
+
+        # res = self.submit(
+        #     QA_Task(
+        #         worker=self.broker[self.get_account(
+        #             account_cookie).broker],
+        #         engine=self.get_account(
+        #             account_cookie).broker,
+        #         event=QA_Event(
+        #             broker=self.broker[self.get_account(
+        #                 account_cookie).broker],
+        #             order_id=order_id
+        #         )
+        #     ),nowait=True)
+
+        return self.order_handler.order_status.loc[account_cookie, realorder_id]
 
     def query_assets(self, account_cookie):
         return self.get_account(account_cookie).assets
@@ -348,7 +460,7 @@ class QA_Market(QA_Trade):
         ))
 
     def query_data(self, broker_name, frequence, market_type, code, start, end=None):
-        self.event_queue.put(
+        self.submit(
             QA_Task(
                 worker=self.broker[broker_name],
                 engine=broker_name,
@@ -385,7 +497,7 @@ class QA_Market(QA_Trade):
     def _trade(self, event):
         "内部函数"
 
-        self.event_queue.put(QA_Task(
+        self.submit(QA_Task(
             worker=self.broker[event.broker_name],
             engine=event.broker_name,
             event=QA_Event(
@@ -403,12 +515,9 @@ class QA_Market(QA_Trade):
         # 向事件线程发送ACCOUNT的SETTLE事件
 
         for account in self.session.values():
-
             if account.running_environment == RUNNING_ENVIRONMENT.TZERO:
-
                 for order in account.close_positions_order:
-
-                    self.event_queue.put(
+                    self.submit(
                         QA_Task(
                             worker=self.broker[account.broker],
                             engine=account.broker,
@@ -418,13 +527,13 @@ class QA_Market(QA_Trade):
                                 callback=self.on_insert_order)))
 
             if account.broker == broker_name:
-                self.event_queue.put(
+                self.submit(
                     QA_Task(
                         worker=account,
                         engine=broker_name,
                         event=QA_Event(
                             event_type=ACCOUNT_EVENT.SETTLE)))
-        self.event_queue.put(QA_Task(
+        self.submit(QA_Task(
             worker=self.broker[broker_name],
             engine=broker_name,
             event=QA_Event(