+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 <!-- TOC -->
 
 - [QUANTAXIS 更新纪要](#quantaxis-更新纪要)
-    - [1.0.69 (unreleased)](#1069-unreleased)
+    - [1.1.0](#110)
     - [1.0.68](#1068)
     - [1.0.67](#1067)
     - [1.0.66](#1066)
@@ -50,9 +50,43 @@
     - [1.0.25](#1025)
 
 <!-- /TOC -->
-## 1.0.69 (unreleased)
+
+## 1.1.0 
+
+[QAData]
+
+1. 修改了采样函数的写法
+2. 修复了tick采样成60min的bug
+
+[QAMARKET]
 
 
+1. 增加对于实盘易的支持
+2. 将一些基础解析字段挪至基类QABroker中
+3. 基于QAMarket创建的broker/order线程全部变成后台线程
+4. QA_OrderHandler 增加持久化 订单/成交单部分
+
+[QAEngine]
+
+1. QAThread 初始化增加 daemon选项, 用于创建守护线程
+
+[QAARP]
+
+1. QA_Account 增加 cancel_order方法 撤单操作
+
+
+[QASU]
+
+1. save_orderhandler.py 增加 QA_SU_save_order/ QA_SU_save_deal 方法
+2. 增加对于运行时出错的容错处理
+
+[QAUtil]
+
+1. QADate_Trade 增加QA_util_get_order_day() 用于获取委托的真实日期
+
+[QAWEB]
+
+1. QAWEB 增加查询股票名称的接口 http://ip:port/marketdata/stock/code?code=xxxxx
 
 ## 1.0.68